requirements.txt
streamlit
google-generativeai
weaviate-client
googletrans==3.1.0a0
pandas
numpy
python-dotenv
langchain
langchain-google-genai
langchain-community
flask
flask-cors
